Fixed Reverse Scope Change Bug.
Esther uncovered an issue with the validation engine while working on
her game tangramsracetwo. This bug caused an issue where if the parent
object was going from game wide to non for its output to the next object
and the next object was already set to game wide, that objects scope
would be set to 0 or nothing, when it should be set to the parents
scope. This was a bug with the andScopeMasks. A new function was
introduced to modify the andScopeMasks appropriately.